Local config menu prep and firmware tinkering
Started to implement a local config menu on the LCD that can be accessed
through the capacitive buttons and uncovered some underlying firmware
changes that needed to be made to enable this. Polling buttons turned
out to not be ideal and in order to properly implement interrupts for
them I had to disable the repeat interrupt and release button interrupt
signals, both of which were not possible with the base Sensorian
firmware. Once I flesh the changes out and tidy them up I shall
contribute them back to the Sensorian GitHub repo.