Updated Threads and CPU temp config menu
Added CPU temp interval to the config menu with some preset options in
there. When the interval is changed it kills the existing thread and
starts a new one, which led me to redesign how threads are started to
allow for maximum code re-use for the rest of the menu options to come.