Finished Config Menu and some bug fixes
Finished the config menu to meet the spec of #2, still more features
that can be added to it as outlined in other outstanding issues. Made it
so the cursor loops back around from bottom to top and back, fixed the
last element of a list not showing and increased reliability of the
accelerometer thread.