Toggle components and System fix
Added booleans to the config file and menu to allow the user to disable
certain sensors as originally intended as well as re-enable them on the
fly through the config menu. This closes #3. Also fixed the issue of not
being able to shutdown or reboot from the config menu when the program
was launched at boot from crontab.  This fixed #13. It also kills the
program gracefully now since things like writing the config file did not
have time to do so in the previous version.