Basic Flask integration
Integrated a test Flask server on its own thread/process to demonstrate
the ability to receive commands from external sources. Would require
them to come from the local network or one with control over port
forwarding given that they are just POSTed, but this would be a simple
solution in most cases when one does not want to utilize a separate,
public facing relay server.