Basic remote config and command implemented
Can now query current config variables with a GET returning a JSON of
them all, or a specific variable on its own URL. A PUT to said URL of
the new value will update it. Currently restricted to the two
orientation variables. Need to add the others. Can send Kill, Shutdown
and Reboot commands as well. Basic Base64 Authentication enabled for the
API but could use improvement. Getting and setting config values moved
to their own functions to be adopted by the rest of the menu options
next.