Testing new install script
New install script to use pip to install python modules, currently just
any version but with the intention of using specific versions in the
future. Also added the Flask modules for remote configuration and
removed WTForms because it isn't used yet, possibly making a return in
the future if I make a web interface for remote config/viewing.