Dependencies change for Pillow
Changed build script to uninstall PIL and Install Pillow 3.2.0 to fix a
memory leak in the display method. This closes #14 but also opens #15.
Also added a check on whether or not to start Flask based on the enabled
variable in the code, in prep for it later being added to the config
file. The new build script needs to be tested on a fresh install.