Setters for UI configs and changed bool_check
Added the remaining setters for the UI-related config file settings.
Also wanted to commit this change for the changes to bool_check. Instead
of returning an int to indicate if it was true, false or neither, it now
returns a 2 item dict that contains a boolean if it was indeed a
boolean, and then the boolean which it was. This allowed me to cut down
the code in the setters significantly seeing as how there are still all
the other sections to add.