Local and API in sync, local menu code cleanup
Now both the local config menu and the API use the same getters and
setters and reboot the threads if necessary using the same code to
ensure whether changing the config locally or remotely the changes are
always in sync with each other. Also moved a lot of locking code for
changing menu elements to set_menu_elements. Could be even further
cleaned up by combining it with the existing change_menu, also making it
easier in the future for possibly allowing for modular lists through the
config. Some minor refactoring as well to supress some of the few
remaining PyCharm warnings.