Fixes regression of local menu button
Removing pin from the arguments of button_event_handler caused the
program to error when receiving a button press, trying to pass the pin
argument to a function that doesn't accept one. Made it check that the
pin indeed came from the buttons just to be safe and make the IDE not
complain about pin being unused anymore rather than simply ignoring it.