Added SocketThread
Added a thread to create and maintain a socket with a public relay
server to receive POSTs from services like IFTTT when the client is in
an uncontrolled networking environment. Also basic IFFTT functionality
to fire off events if the up or down button are pressed when not in the
menu.