New light demo and door magnet demo
Changed the light demo to change the brightness of a local Philips Hue
light to compensate for another going brighter or darker like the sun.
Added a demo to act as a door security system that can determine if a
door is open or closed if placed on it with a magnet nearby on the
frame. In order to do this a thread and function to monitor the
magnetometer was created. These demos should be split off into separate
code if possible.