Preliminary support for other Hats
Added options to the config file to enable or disable the use of hats
such as the Sensorian Shield. Sense HAT and MATRIX Creator. Enabling
them disables the use of individual breadboarded sensors which will be
added later. Defaults to Sensorian. Next up is supporting the other two
hats mentioned above!