Added basic Sense HAT support
Can now select the Sense HAT for use which will collect the temperature,
pressure, accelerometer and magnetometer values. Changed date functions
to construct a standard python datetime object, with one being
constructed with the time of now if the Sensorian Shield real time clock
is not used. Should add support for the Sense HAT display, joystick, and
humidity etc.